Refactored the file structure of package for distributed_cli
